+++ src/main/java/lotto/model/LottoResult.java
@@ -1,8 +1,4 @@
-package step2.model;
-
-import static step2.model.LottoBonusResult.MATCHED;
-import static step2.model.LottoBonusResult.NOTHING;
-import static step2.model.LottoBonusResult.NOT_MATCHED;
+package lotto.model;
 
 import java.util.Collections;
 import java.util.Map;
@@ -12,14 +8,14 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 public enum LottoResult {
-    NONE(LottoMatchCount.newInstance(0, NOTHING), 0L),
-    ONE(LottoMatchCount.newInstance(1, NOTHING), 0L),
-    TWO(LottoMatchCount.newInstance(2, NOTHING), 0L),
-    THREE(LottoMatchCount.newInstance(3, NOTHING), 5_000L),
-    FOUR(LottoMatchCount.newInstance(4, NOTHING), 50_000L),
-    FIVE(LottoMatchCount.newInstance(5, NOT_MATCHED ), 1_500_000L),
-    FIVE_WITH_BONUS(LottoMatchCount.newInstance(5, MATCHED), 30_000_000L),
-    SIX(LottoMatchCount.newInstance(6, NOTHING), 2_000_000_000L);
+    NONE(LottoMatchCount.newInstance(0, LottoBonusResult.NOTHING), 0L),
+    ONE(LottoMatchCount.newInstance(1, LottoBonusResult.NOTHING), 0L),
+    TWO(LottoMatchCount.newInstance(2, LottoBonusResult.NOTHING), 0L),
+    THREE(LottoMatchCount.newInstance(3, LottoBonusResult.NOTHING), 5_000L),
+    FOUR(LottoMatchCount.newInstance(4, LottoBonusResult.NOTHING), 50_000L),
+    FIVE(LottoMatchCount.newInstance(5, LottoBonusResult.NOT_MATCHED ), 1_500_000L),
+    FIVE_WITH_BONUS(LottoMatchCount.newInstance(5, LottoBonusResult.MATCHED), 30_000_000L),
+    SIX(LottoMatchCount.newInstance(6, LottoBonusResult.NOTHING), 2_000_000_000L);
 
     private static final Map<LottoMatchCount, LottoResult> LOTTO_RESULTS =
         Collections.unmodifiableMap(Stream.of(values())
