+++ src/main/java/lotto/model/LottoStore.java
@@ -1,4 +1,4 @@
-package step2.model;
+package lotto.model;
 
 import java.util.List;
 import java.util.stream.Collectors;
