+++ src/test/java/step2/model/LottoStoreTest.java
@@ -3,6 +3,9 @@ package step2.model;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.stream.Stream;
+import lotto.model.LottoStore;
+import lotto.model.Payment;
+import lotto.model.PurchasedLottoTickets;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
