+++ src/test/java/step2/model/PaymentTest.java
@@ -3,6 +3,7 @@ package step2.model;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 import java.util.stream.Stream;
+import lotto.model.Payment;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
