+++ src/test/java/step2/model/WinningLottoTicketTest.java
@@ -6,6 +6,8 @@ import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
+import lotto.model.LottoNumber;
+import lotto.model.WinningLottoTicket;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
